chore: Move useDebounce to utils
