+++ src/app/(home)/search/useSearch.ts
@@ -1,14 +1,6 @@
-import { IResult } from "@/types/data";
 import { useEffect, useState } from "react";
-
-function useDebounce<T>(value: T, delay: number): T {
-  const [debouncedValue, setDebouncedValue] = useState<T>(value);
-  useEffect(() => {
-    const handler = setTimeout(() => setDebouncedValue(value), delay);
-    return () => clearTimeout(handler);
-  }, [value, delay]);
-  return debouncedValue;
-}
+import { IResult } from "@/types/data";
+import { useDebounce } from "@/utils/hooks";
 
 function useSearch(q?: string) {
   const [results, setResults] = useState<IResult[]>();
